+++ client/src/assets/css/notebook.css
