+++ client/src/components/DeleteSheetButton.tsx
