+++ server/src/datasources/user.ts
@@ -1,6 +1,6 @@
 import { DataSource, DataSourceConfig } from 'apollo-datasource';
 import isEmail from 'isemail';
-import Notebook from './cookbook';
+import Notebook from './notebook';
 import { db, dbUser, UserModel } from '../store';
 import { Context } from '..';
 import UserGQL from '../graphql_models/userGQL';
