+++ server/src/index.ts
@@ -7,10 +7,10 @@ import resolvers from './resolvers';
 import { dbUser, UserModel } from './store';
 import User from './datasources/user';
 import Avatar from './datasources/avatar';
-import Todo from './datasources/ingredient';
-import Sheet from './datasources/recipe';
+import Todo from './datasources/todo';
+import Sheet from './datasources/sheet';
 import path from 'path';
-import Notebook from './datasources/cookbook';
+import Notebook from './datasources/notebook';
 
 interface MyContext {
   user: UserModel | null;
