+++ server/src/resolvers.ts
@@ -11,9 +11,9 @@ import {
 } from './schema';
 import UserGQL from './graphql_models/userGQL';
 import AvatarGQL from './graphql_models/AvatarGQL';
-import SheetGQL from './graphql_models/recipeGQL';
-import TodoGQL from './graphql_models/ingredientGQL';
-import NotebookGQL from './graphql_models/cookbookGQL';
+import SheetGQL from './graphql_models/sheetGQL';
+import TodoGQL from './graphql_models/todoGQL';
+import NotebookGQL from './graphql_models/notebookGQL';
 
 const resolvers = {
   Query: {
