+++ server/src/schema.ts
@@ -1,9 +1,9 @@
 import { gql } from 'apollo-server';
 import UserGQL from './graphql_models/userGQL';
 import AvatarGQL from './graphql_models/AvatarGQL';
-import NotebookGQL from './graphql_models/cookbookGQL';
-import SheetGQL from './graphql_models/recipeGQL';
-import TodoGQL from './graphql_models/ingredientGQL';
+import NotebookGQL from './graphql_models/notebookGQL';
+import SheetGQL from './graphql_models/sheetGQL';
+import TodoGQL from './graphql_models/todoGQL';
 
 export interface UserResponseGQL {
   success: boolean;
